edits

+++ 0x13-more_singly_linked_lists/102-free_listint_safe.c
@@ -14,7 +14,7 @@ size_t free_listint_safe(listint_t **h)
 {
 	listint_t *tmp, *actual;
 	size_t i;
-	ptrdiff_t rest;
+	int rest;
 
 	i = 0, actual = *h;
 	while (actual)
